Support BIGINT cutoff_rank to prevent overflow and force drop table cascade on seeder refresh

+++ backend/schema_postgres.sql
@@ -72,7 +72,7 @@ CREATE TABLE IF NOT EXISTS cutoffs (
     course_id INTEGER NOT NULL,
     round INTEGER NOT NULL,
     category TEXT NOT NULL,
-    cutoff_rank INTEGER NOT NULL,
+    cutoff_rank BIGINT NOT NULL,
     year INTEGER DEFAULT 2025,
     FOREIGN KEY(course_id) REFERENCES courses(id) ON DELETE CASCADE
 );
